backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
motor==3.5.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
PyJWT==2.9.0
passlib==1.7.4
bcrypt==4.2.0
email-validator==2.2.0
fastapi-mail==1.4.1
Pillow==10.4.0
qrcode==7.4.2
python-multipart==0.0.9
